move plotting utils from casie to common
