+++ bindings/pydairlib/analysis/log_plotter_cassie.py
@@ -31,37 +31,13 @@ def main():
 
     t_x_slice = slice(robot_output['t_x'].size)
 
-    plot_floating_base_positions(robot_output, q_names, t_x_slice)
-    plot_joint_positions(robot_output, q_names, t_x_slice, floating_base=True)
+    cassie_plotting_utils.plot_floating_base_positions(
+        robot_output, q_names, t_x_slice)
+    cassie_plotting_utils.plot_joint_positions(
+        robot_output, q_names, t_x_slice, floating_base=True)
 
     plt.show()
 
-
-def plot_floating_base_positions(robot_output, q_names, time_slice):
-    cassie_plotting_utils.make_plot(
-        robot_output,                       # data dict
-        't_x',                              # time channel
-        time_slice,
-        ['q'],                              # key to plot
-        {'q': slice(0, 7)},                 # slice of 'q' to plot
-        {'q': q_names[:7]},                 # legend entries
-        {'xlabel': 'Time',
-         'ylabel': 'Position',
-         'title': 'Floating Base Positions'})
-
-
-def plot_joint_positions(robot_output, q_names, time_slice, floating_base=True):
-    q_slice = slice(7, len(q_names)) if floating_base else slice(len(q_names))
-    cassie_plotting_utils.make_plot(
-        robot_output,                       # data dict
-        't_x',                              # time channel
-        time_slice,
-        ['q'],                              # key to plot
-        {'q': q_slice},                     # slice of 'q' to plot
-        {'q': q_names[q_slice]},            # legend entries
-        {'xlabel': 'Time',
-         'ylabel': 'Joint Angle (rad)',
-         'title': 'Joint Positions'})
 
 if __name__ == '__main__':
     main()
